file structure restructure next.. authentication
