+++ app/app.py
@@ -1,7 +1,9 @@
 from flask import Flask
 from flask_migrate import Migrate
 from flask_cors import CORS
-from main import main_bp
+# from main import main_bp
+from userroute import user_bp
+from interests import interest_bp
 import os
 
 
@@ -17,8 +19,10 @@ def create_app():
     jwt.init_app(app)
     bcrypt.init_app(app)
     migrate = Migrate(app, db)
-    app.register_blueprint(main_bp)
+    # app.register_blueprint(main_bp)
     app.register_blueprint(auth_bp)
+    app.register_blueprint(user_bp)
+    app.register_blueprint(interest_bp)
     CORS(app, resources={r"*": {"origins": "*"}})
